requirements.txt
requests
markdown
beautifulsoup4
markdownify
python-dotenv
